more progress, also fixed formatting issue and missing back scatt cross in job caching

+++ src/cc_hex_mod.f90
@@ -35,7 +35,7 @@ module cc_hex_mod
         INTEGER :: O1 ! HB 23/1/23 offset test
         
         ! HB integration with abt
-        integer, dimension(:), allocatable, intent(out) :: apertures ! taken as parents parent facets
+        integer(8), dimension(:), allocatable, intent(out) :: apertures ! taken as parents parent facets
         real(8), dimension(:,:), allocatable, intent(out) :: vertices
         integer(8), dimension(:,:), allocatable, intent(out) :: face_ids
         integer face_counter, vert_counter ! counter for tracking number of faces
